refactored container builder
