+++ TagCloud/TagCloud/TagCloud.csproj
@@ -72,6 +72,7 @@
     <Compile Include="App\UiActions\FontSettingsAction.cs" />
     <Compile Include="App\UiActions\BoringWordsListAction.cs" />
     <Compile Include="App\UiActions\FilterListAction.cs" />
+    <Compile Include="App\Builder.cs" />
     <Compile Include="Infrastructure\BoringWord.cs" />
     <Compile Include="Infrastructure\Interfaces\ICheckable.cs" />
     <Compile Include="Infrastructure\Interfaces\IImageHolder.cs" />
